NPQ loop 300 seconds
Added a loop that calculates the photosynthesis rate for 300 iterations after the light transition. For the first 50 seconds photosynthesis is calculated for every second. After 50 seconds, every 10 seconds the photosynthesis is calculated to decrease computational power. This is done because after 50 seconds, no large change is seen in any of the output variables.